@@ -7,7 +7,7 @@
 
 #include "Logic.hpp"
 #include "ir/operations/OpType.hpp"
-#include "utils.hpp"
+#include "sc/utils.hpp"
 
 #include <cstddef>
 #include <plog/Log.h>
