@@ -11,7 +11,6 @@
 #include "ir/operations/AodOperation.hpp"
 #include "ir/operations/OpType.hpp"
 #include "ir/operations/Operation.hpp"
-#include "nlohmann/json.hpp"
 
 #include <algorithm>
 #include <cstddef>
@@ -19,7 +18,7 @@
 #include <exception>
 #include <fstream>
 #include <map>
-#include <nlohmann/json_fwd.hpp>
+#include <nlohmann/json.hpp>
 #include <set>
 #include <stdexcept>
 #include <string>
@@ -28,7 +27,7 @@
 namespace na {
 
 void NeutralAtomArchitecture::loadJson(const std::string& filename) {
-  nlohmann::json jsonData;
+  nlohmann::basic_json<> jsonData;
   std::ifstream architectureFile(filename);
 
   if (!architectureFile.is_open()) {
@@ -39,7 +38,7 @@ void NeutralAtomArchitecture::loadJson(const std::string& filename) {
     architectureFile.close();
 
     // Load properties
-    nlohmann::json jsonDataProperties = jsonData["properties"];
+    nlohmann::basic_json<> jsonDataProperties = jsonData["properties"];
     this->properties = Properties(
         jsonDataProperties["nRows"], jsonDataProperties["nColumns"],
         jsonDataProperties["nAods"], jsonDataProperties["nAodCoordinates"],
@@ -49,7 +48,7 @@ void NeutralAtomArchitecture::loadJson(const std::string& filename) {
         jsonDataProperties["minimalAodDistance"]);
 
     // Load parameters
-    const nlohmann::json jsonDataParameters = jsonData["parameters"];
+    const nlohmann::basic_json<> jsonDataParameters = jsonData["parameters"];
     this->parameters = Parameters();
     this->parameters.nQubits = jsonDataParameters["nQubits"];
 
