@@ -3,9 +3,9 @@
 // See README.md or go to https://github.com/cda-tum/qmap for more information.
 //
 
-#include "Architecture.hpp"
-#include "Configuration.hpp"
 #include "ir/operations/OpType.hpp"
+#include "na/Architecture.hpp"
+#include "na/Configuration.hpp"
 
 #include <cstdint>
 #include <cstdio>
